Add a nova-vms-on-host tool.
We had a host die last week and it's really annoying to iterate all the
VMs to get customer contact info. That inspired this tool. Given a
hypervisor, get all customer names and emails (pasteable) so that you
can let the customers know about a potential issue. Also useful for
maintenance.

Change-Id: I783a942755d9c61a33a5f32ded10391db56ba124