A script to count the number of routers by host
This can be used during maintenance to ensure all routers have moved or
when you believe the number of routers is not balanced on your nodes. It
does not however count FIPs per router or factor traffic into the count,
so it can only be used to get a general idea of the balance.

Change-Id: I5003c578d338d34d09608a979eb469d0465c0af3