SuperHAProxy
SuperHAProxy is a script around HAProxy to add support for live
migration and easy handling of multiple instances. Live migration is
added using the CRIU system. Very lightweight containers are used to
isolate the proxy instances from each other to enable live migration
to efficiently transfer the proxies and to ensure instances can be
running different versions to support proper rolling upgrades.

This script can be used stand alone, or as a building block to
provide multiserver proxy pools that are high performance and support
a workflow where the servers can be rolling upgraded live without
connection loss.

Change-Id: Id3150c5d5de6c3f5552f27f79cbcf0debe8adb1f