Remove hardcoded volume point and network device and tenant network
This patch set fixed the problems that the script was using hardcoded
device names such as eth0, /dev/vdb. The fix will detect the network
devices where possible or make it configurable. In this patch, the
tenant network is no longer required for ansible lampstack test. If
a cloud does not support tenant network, simply leave the two network
parameters empty.

Change-Id: Ic4bcd07caa7f7a27f7cb520fb3302fb547f085f0