Rescan scsi bus before using volume
Depending on the hypervisor and disk adapter used on the guest OS,
it may be necessary to rescan the SCSI bus before a Cinder volume
that has recently been attached is seen by the OS.  This patch installs
the scsitools package and rescans the bus prior to partitioning new
space from the attached volume.  It is harmless for
hypervisor/adapter/guestOS combinations that don't need it (other than
taking a tiny amount of disk space and a few seconds), and makes the
workload more portable for those that do.

Change-Id: Ibb33ec359343a7093cf3f6aa0d0a7e9086b9a55b