Make lampstack work in config drive environments
Currently, the lampstack workload assumes that either you're using the
nova metadata service or that config drive is enabled globally in
your nova.conf file: there's no option to enable it on a per-environment
basis.  This patch adds the ability to specify whether config_drive
should be used or not in the app_env variables, with a default value
set to "no" (which means that the new variable isn't required and that
if a user doesn't set it, he or she will get the same behavior as before
this patch).

Change-Id: I717bf5a37f98da79141af70df90feb7fd7b3f8c7