Update apt cache before installing scsitools
A few folks reported that installing scsitools was throwning an error
for them because the apt cache on the system hadn't been initialized.
This patch ensures that we update the apt cache before attempting to
install scsitools.

The patch also makes the workload run faster overall: previously,
we updated the apt cache each time we installed a package.  In this
patch, we break out the cache update into a separate task that runs
before any software is installed, thus eliminating several
time-consuming calls to "apt-get update".

Finally, this patch also makes the workload slightly more portable:
it replaces apt calls with package calls.  Adding a new guest OS
family should just require adding a new cache update block (if
necessary for that OS) with an appropriate when statement.

Change-Id: I30ffbf2da13f22475f920328d84e64460834bcdb